+++ crypto/fipsmodule/CMakeLists.txt
@@ -33,7 +33,6 @@ if(ARCH STREQUAL "x86_64")
     aesni-x86_64.${ASM_EXT}
     ghash-ssse3-x86_64.${ASM_EXT}
     ghash-x86_64.${ASM_EXT}
-    md5-x86_64.${ASM_EXT}
     p256-x86_64-asm.${ASM_EXT}
     p256_beeu-x86_64-asm.${ASM_EXT}
     rdrand-x86_64.${ASM_EXT}
@@ -59,7 +58,6 @@ if(ARCH STREQUAL "x86")
     co-586.${ASM_EXT}
     ghash-ssse3-x86.${ASM_EXT}
     ghash-x86.${ASM_EXT}
-    md5-586.${ASM_EXT}
     sha1-586.${ASM_EXT}
     sha256-586.${ASM_EXT}
     sha512-586.${ASM_EXT}
@@ -97,7 +95,6 @@ if(ARCH STREQUAL "aarch64")
     ghash-neon-armv8.${ASM_EXT}
     ghashv8-armx.${ASM_EXT}
     keccak1600-armv8.${ASM_EXT}
-    md5-armv8.${ASM_EXT}
     p256-armv8-asm.${ASM_EXT}
     p256_beeu-armv8-asm.${ASM_EXT}
     rndr-armv8.${ASM_EXT}
@@ -142,9 +139,6 @@ if(PERL_EXECUTABLE)
   perlasm(ghash-x86_64.${ASM_EXT} modes/asm/ghash-x86_64.pl)
   perlasm(ghash-x86.${ASM_EXT} modes/asm/ghash-x86.pl)
   perlasm(keccak1600-armv8.${ASM_EXT} sha/asm/keccak1600-armv8.pl)
-  perlasm(md5-586.${ASM_EXT} md5/asm/md5-586.pl)
-  perlasm(md5-armv8.${ASM_EXT} md5/asm/md5-armv8.pl)
-  perlasm(md5-x86_64.${ASM_EXT} md5/asm/md5-x86_64.pl)
   perlasm(p256-x86_64-asm.${ASM_EXT} ec/asm/p256-x86_64-asm.pl)
   perlasm(p256_beeu-x86_64-asm.${ASM_EXT} ec/asm/p256_beeu-x86_64-asm.pl)
   perlasm(p256-armv8-asm.${ASM_EXT} ec/asm/p256-armv8-asm.pl)
